feat : GBN 라우터만 구성 / 메인 와이어프레임 잡기
