+++ src/app/community/page.tsx
@@ -1,0 +1,3 @@
+export default function CommunityPage() {
+  return <>커뮤니티</>;
+}
