+++ src/app/interested-certifications/page.tsx
@@ -1,0 +1,3 @@
+export default function InterestedCertificationsPage() {
+  return <>관심자격증</>;
+}
