+++ src/app/mypage/page.tsx
@@ -1,0 +1,3 @@
+export default function MypagePage() {
+  return <>마이페이지</>;
+}
